+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Accounting tool: correct table formatting for Income in PDF report.
 - Coinbase parser: strip currency symbols, brackets and commas from fiat values.
 - Conversion tool: Interest should be mapped to Income not LoanInterest for RECAP output format.
+- Price tool: CryptoCompare API rate limit reached returns 200 OK. ([#381](https://github.com/BittyTax/BittyTax/issues/381))
 ### Added
 - Accounting tool: new PDF report format.
 - Config: added classic_report parameter to switch to legacy PDF report format.